// types/game/singleGame.ts
interface Store {
  id: number
  url: string
  store: {
    id: number
    name: string
    slug: string
    domain: string
    games_count: number
    image_background: string
  }
}

interface Genre {
  id: number
  name: string
  slug: string
  language: string
  games_count: number
  image_background: string
}

interface Rating {
  id: number
  title: string
  count: number
  percent: number
}
// ...
interface AddedByStatus {
  yet: number
  owned: number
  beaten: number
  toplay: number
  dropped: number
  playing: number
}

interface PlatformInfo {
  id: number
  name: string
  slug: string
  image: string | null
  year_start: number | null
  year_end: number | null
  games_count: number
  image_background: string
}

interface Store {
  id: number
  store: {
    id: number
    name: string
    slug: string
    domain: string
    games_count: number
    image_background: string
  }
}

interface Genre {
  id: number
  name: string
  slug: string
  games_count: number
  image_background: string
}

interface Tag {
  id: number
  name: string
  slug: string
  language: string
  games_count: number
  image_background: string
}

interface ShortScreenshot {
  id: number
  image: string
}

export interface GameDlc {
  id: number
  slug: string
  name: string
  released: string
  tba: boolean
  background_image: string
  rating: number
  rating_top: number
  ratings: Rating[]
  ratings_count: number
  reviews_text_count: number
  added: number
  added_by_status: AddedByStatus
  metacritic: number
  playtime: number
  suggestions_count: number
  updated: string
  user_game: any
  reviews_count: number
  saturated_color: string
  dominant_color: string
  platforms: {
    platform: PlatformInfo
    released_at: string
    requirements_en: any
    requirements_ru: any
  }[]
  parent_platforms: { platform: PlatformInfo }[]
  genres: Genre[]
  stores: Store[]
  clip: any
  tags: Tag[]
  esrb_rating: { id: number; name: string; slug: string }
  short_screenshots: ShortScreenshot[]
}